refactor: cargo fmt code base
