+++ boards/esp32c3_ble/build.rs
@@ -1,15 +1,3 @@
-//! This build script copies the `memory.x` file from the crate root into
-//! a directory where the linker can always find it at build time.
-//! For many projects this is optional, as the linker always searches the
-//! project root directory -- wherever `Cargo.toml` is. However, if you
-//! are using a workspace or have a more complicated build setup, this
-//! build script becomes required. Additionally, by requesting that
-//! Cargo re-run the build script whenever `memory.x` is changed,
-//! updating `memory.x` ensures a rebuild of the application with the
-//! new memory settings.
-//!
-//! The build script also sets the linker flags to tell it which link script to use.
-
 use const_gen::*;
 use std::fs::File;
 use std::io::Read;
@@ -21,29 +9,8 @@ fn main() {
     // Generate vial config at the root of project
     generate_vial_config();
 
+    // ESP IDE system env
     embuild::espidf::sysenv::output();
-    // embuild::build::CfgArgs::output_propagated("ESP_IDF").unwrap();
-    // embuild::build::LinkArgs::output_propagated("ESP_IDF").unwrap();
-    // Put `memory.x` in our output directory and ensure it's
-    // on the linker search path.
-    // let out = &PathBuf::from(env::var_os("OUT_DIR").unwrap());
-    // File::create(out.join("memory.x"))
-    //     .unwrap()
-    //     .write_all(include_bytes!("memory.x"))
-    //     .unwrap();
-    // println!("cargo:rustc-link-search={}", out.display());
-
-    // By default, Cargo will re-run a build script whenever
-    // any file in the project changes. By specifying `memory.x`
-    // here, we ensure the build script is only re-run when
-    // `memory.x` is changed.
-    // println!("cargo:rerun-if-changed=memory.x");
-
-    // Specify linker arguments.
-
-    // Set the linker script to the one provided by espressif.
-    // println!("cargo:rustc-link-arg=-Tlinkall.x");
-    // println!("cargo:rustc-link-arg=-Trom_functions.x");
 
     // Set the extra linker script from defmt
     println!("cargo:rustc-link-arg=-Tdefmt.x");
