+++ rust-toolchain.toml
@@ -1,5 +1,5 @@
 [toolchain]
-channel = "nightly"
+channel = "1.75"
 components = ["rust-src", "rustfmt", "llvm-tools"]
 targets = [
     "thumbv7em-none-eabi",
